requirements.txt
openai
python-dotenv
pydantic
fastapi
requests
bs4